+++ Classes/Growler.m
@@ -1,12 +1,15 @@
-//
-//  Growler.m
-//  Hermes
-//
-//  Created by Alex Crichton on 10/2/11.
-//
+/**
+ * @file Growler.h
+ * @brief Growl integration for the rest of Hermes
+ *
+ * Provides unified access to displaying notifications for different kinds
+ * of events without having to deal with Growl directly.
+ */
 
 #import <Growl/Growl.h>
+
 #import "Growler.h"
+#import "PreferencesController.h"
 
 Growler *growler = nil;
 
@@ -24,15 +27,21 @@ Growler *growler = nil;
   [GrowlApplicationBridge setGrowlDelegate:nil];
 }
 
-+ (void) growl:(Song*)song withImage:(NSImage*)image {
++ (void) growl:(Song*)song withImage:(NSImage*)image isNew:(BOOL)n {
   if (growler == nil) {
     return;
   }
 
-  [growler growl:song withImage:image];
+  [growler growl:song withImage:image isNew:n];
 }
 
-- (void) growl:(Song*)song withImage:(NSImage*)image {
+- (void) growl:(Song*)song withImage:(NSImage*)image isNew:(BOOL)n {
+  NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
+  if ((n && ![defaults boolForKey:PLEASE_GROWL_NEW]) ||
+      (!n && ![defaults boolForKey:PLEASE_GROWL_PLAY])) {
+    return;
+  }
+
   NSString *title = [song title];
   NSString *description = [song artist];
   description = [description stringByAppendingString:@" - "];
@@ -44,7 +53,7 @@ Growler *growler = nil;
      something that's plist-encodable */
   [GrowlApplicationBridge notifyWithTitle:title
                               description:description
-                         notificationName:@"hermes-song"
+                         notificationName:n ? @"hermes-song" : @"hermes-play"
                                  iconData:[image TIFFRepresentation]
                                  priority:0
                                  isSticky:NO
@@ -60,8 +69,14 @@ Growler *growler = nil;
   NSMutableDictionary *dict = [NSMutableDictionary dictionary];
   NSMutableArray *notifications = [NSMutableArray array];
   [notifications addObject:@"hermes-song"];
+  [notifications addObject:@"hermes-play"];
   [dict setObject:notifications forKey:GROWL_NOTIFICATIONS_ALL];
   [dict setObject:notifications forKey:GROWL_NOTIFICATIONS_DEFAULT];
+
+  NSMutableDictionary *human_names = [NSMutableDictionary dictionary];
+  [human_names setObject:@"hermes-song" forKey:@"New Songs"];
+  [human_names setObject:@"hermes-play" forKey:@"Play/pause Events"];
+  [dict setObject:human_names forKey:GROWL_NOTIFICATIONS_HUMAN_READABLE_NAMES];
   return dict;
 }