+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 1.0.21 (unreleased)
+
+[Full changelog](https://github.com/alexcrichton/hermes/compare/v1.0.20...master)
+
+* [added] Fine-grained control for which growl notifications are received,
+          configurable through Hermes' preferences, or also through Growl's
+
 # Version 1.0.20 (2/23/12)
 
 [Full changelog](https://github.com/alexcrichton/hermes/compare/v1.0.19...v1.0.20)